+++ crates/forge/src/warn.rs
@@ -4,6 +4,7 @@ use forge_runner::backtrace::is_backtrace_enabled;
 use forge_runner::package_tests::with_config_resolved::TestTargetWithResolvedConfig;
 use foundry_ui::UI;
 use foundry_ui::components::warning::WarningMessage;
+use indoc::formatdoc;
 use scarb_api::{ScarbCommand, package_matches_version_requirement};
 use scarb_metadata::Metadata;
 use semver::{Comparator, Op, Version, VersionReq};
@@ -184,7 +185,7 @@ pub fn warn_if_snforge_std_does_not_match_package_version(
     Ok(())
 }
 
-// TODO(#3272)
+// TODO(#3679): Remove this function when we decide to bump minimal scarb version to 2.12.
 pub(crate) fn warn_if_backtrace_without_panic_hint(scarb_metadata: &Metadata, ui: &UI) {
     if is_backtrace_enabled() {
         let is_panic_backtrace_set = scarb_metadata
@@ -202,10 +203,18 @@ pub(crate) fn warn_if_backtrace_without_panic_hint(scarb_metadata: &Metadata, ui
             });
 
         if !is_panic_backtrace_set {
-            ui.println(
-                &WarningMessage::new("To get accurate backtrace results, it is required to use the configuration available in the latest Cairo version. \
-                For more details, please visit https://foundry-rs.github.io/starknet-foundry/snforge-advanced-features/backtrace.html"
-            ));
+            let message = formatdoc! {
+                "Scarb version should be 2.12 or higher and `Scarb.toml` should have the following Cairo compiler configuration to get accurate backtrace results:
+
+                [profile.{profile}.cairo]
+                unstable-add-statements-functions-debug-info = true
+                unstable-add-statements-code-locations-debug-info = true
+                panic-backtrace = true # only for scarb 2.12 or higher
+                ... other entries ...
+                ",
+                profile = scarb_metadata.current_profile
+            };
+            ui.println(&WarningMessage::new(message));
         }
     }
 }